Move lexer's string table into Identifier and clean up imports
