+++ src/e2ir.c
@@ -16,7 +16,6 @@
 #include        "port.h"
 #include        "target.h"
 
-#include        "lexer.h"
 #include        "expression.h"
 #include        "mtype.h"
 #include        "dsymbol.h"
