+++ src/idgen.c
@@ -398,7 +398,6 @@ int main()
         fprintf(fp, "// File generated by idgen.c\n");
         fprintf(fp, "#include \"id.h\"\n");
         fprintf(fp, "#include \"identifier.h\"\n");
-        fprintf(fp, "#include \"lexer.h\"\n");
 
         for (i = 0; i < sizeof(msgtable) / sizeof(msgtable[0]); i++)
         {   const char *id = msgtable[i].ident;
@@ -418,7 +417,7 @@ int main()
 
             if (!p)
                 p = id;
-            fprintf(fp,"    %s = Lexer::idPool(\"%s\");\n", id, p);
+            fprintf(fp,"    %s = Identifier::idPool(\"%s\");\n", id, p);
         }
 
         fprintf(fp, "}\n");
