+++ src/traits.c
@@ -737,7 +737,7 @@ Expression *semanticTraits(TraitsExp *e, Scope *sc)
             e->error("string must be chars");
             goto Lfalse;
         }
-        Identifier *id = Lexer::idPool((char *)se->string);
+        Identifier *id = Identifier::idPool((char *)se->string);
 
         /* Prefer dsymbol, because it might need some runtime contexts.
          */
